Extract package singleton class for easier testing
